requirements.txt
discord.py
python-dotenv
python-nmap
aiohttp